Moved scope step
