+++ docs/spec/src/access/index.ts
@@ -27,60 +27,4 @@ export const accessSpec: StepSpec<IntegrationConfig>[] = [
     dependsOn: ['fetch-account'],
     implemented: true,
   },
-  {
-    /**
-     * ENDPOINT: https://localhost/api/v1/app_scopes
-     * PATTERN: Fetch Entities
-     */
-    id: 'fetch-scopes',
-    name: 'Fetch Scopes',
-    entities: [
-      {
-        resourceName: 'Scope',
-        _type: 'csw_scope',
-        _class: ['Group'],
-      },
-    ],
-    relationships: [],
-    dependsOn: ['fetch-users'],
-    implemented: true,
-  },
-  {
-    /**
-     * n/a
-     * PATTERN: Build Child Relationships
-     */
-    id: 'build-scope-relationships',
-    name: 'Build Scope -> Scope Relationships',
-    entities: [],
-    relationships: [
-      {
-        _type: 'csw_scope_has_scope',
-        sourceType: 'csw_scope',
-        _class: RelationshipClass.HAS,
-        targetType: 'csw_scope',
-      },
-    ],
-    dependsOn: ['fetch-scopes'],
-    implemented: true,
-  },
-  {
-    /**
-     * n/a
-     * PATTERN: Build Child Relationships
-     */
-    id: 'build-scope-user-relationships',
-    name: 'Build Scope -> User Relationships',
-    entities: [],
-    relationships: [
-      {
-        _type: 'csw_scope_assigned_user',
-        sourceType: 'csw_scope',
-        _class: RelationshipClass.ASSIGNED,
-        targetType: 'csw_user',
-      },
-    ],
-    dependsOn: ['fetch-scopes'],
-    implemented: true,
-  },
 ];
