+++ src/__recordings__/client-id-auth-error_233254982/recording.har
@@ -33,7 +33,7 @@
             {
               "_fromType": "array",
               "name": "timestamp",
-              "value": "2022-06-13T20:11:34+0000"
+              "value": "2022-06-14T15:40:29+0000"
             },
             {
               "_fromType": "array",
@@ -85,35 +85,35 @@
             },
             {
               "name": "date",
-              "value": "Mon, 13 Jun 2022 20:11:35 GMT"
-            },
-            {
-              "name": "x-zen-fury",
-              "value": "1cf02e839e83092aa874f5fcddd8dde9e96431c8"
-            },
-            {
-              "name": "x-xss-protection",
-              "value": "1; mode=block"
+              "value": "Tue, 14 Jun 2022 15:40:30 GMT"
             },
             {
               "name": "server",
               "value": "ZENEDGE"
             },
             {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
               "name": "x-cache-status",
               "value": "NOTCACHED"
             },
             {
-              "name": "content-length",
-              "value": "20"
+              "name": "x-zen-fury",
+              "value": "140aedb1e8edf0d19f13118487c1d20c6b10326d"
+            },
+            {
+              "name": "x-content-type-options",
+              "value": "nosniff"
+            },
+            {
+              "name": "x-xss-protection",
+              "value": "1; mode=block"
             },
             {
               "name": "strict-transport-security",
               "value": "max-age=31536000; includeSubDomains; preload;"
+            },
+            {
+              "name": "content-length",
+              "value": "20"
             },
             {
               "name": "x-cdn",
@@ -126,8 +126,8 @@
           "status": 401,
           "statusText": "Unauthorized"
         },
-        "startedDateTime": "2022-06-13T20:11:34.177Z",
-        "time": 627,
+        "startedDateTime": "2022-06-14T15:40:29.555Z",
+        "time": 460,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -135,7 +135,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 627
+          "wait": 460
         }
       }
     ],
