+++ src/__recordings__/client-secret-auth-error_1001665309/recording.har
@@ -33,7 +33,7 @@
             {
               "_fromType": "array",
               "name": "timestamp",
-              "value": "2022-06-13T20:11:34+0000"
+              "value": "2022-06-14T15:40:30+0000"
             },
             {
               "_fromType": "array",
@@ -84,36 +84,36 @@
               "value": "close"
             },
             {
-              "name": "server",
-              "value": "ZENEDGE"
-            },
-            {
-              "name": "date",
-              "value": "Mon, 13 Jun 2022 20:11:35 GMT"
+              "name": "content-length",
+              "value": "16"
             },
             {
               "name": "x-cache-status",
               "value": "NOTCACHED"
             },
             {
+              "name": "server",
+              "value": "ZENEDGE"
+            },
+            {
+              "name": "x-zen-fury",
+              "value": "7bac94a4d67500d68203ff420f7ce9ec52eaf6b1"
+            },
+            {
               "name": "x-content-type-options",
               "value": "nosniff"
             },
             {
-              "name": "x-zen-fury",
-              "value": "ded44b41ce7bf12bdf9cc55558a0d50ec18853f9"
-            },
-            {
-              "name": "content-length",
-              "value": "16"
-            },
-            {
-              "name": "strict-transport-security",
-              "value": "max-age=31536000; includeSubDomains; preload;"
+              "name": "date",
+              "value": "Tue, 14 Jun 2022 15:40:30 GMT"
             },
             {
               "name": "x-xss-protection",
               "value": "1; mode=block"
+            },
+            {
+              "name": "strict-transport-security",
+              "value": "max-age=31536000; includeSubDomains; preload;"
             },
             {
               "name": "x-cdn",
@@ -126,8 +126,8 @@
           "status": 403,
           "statusText": "Forbidden"
         },
-        "startedDateTime": "2022-06-13T20:11:34.821Z",
-        "time": 363,
+        "startedDateTime": "2022-06-14T15:40:30.027Z",
+        "time": 354,
         "timings": {
           "blocked": -1,
           "connect": -1,
@@ -135,7 +135,7 @@
           "receive": 0,
           "send": 0,
           "ssl": -1,
-          "wait": 363
+          "wait": 354
         }
       }
     ],
