+++ src/steps/access/converter.ts
@@ -7,7 +7,7 @@ import {
 } from '@jupiterone/integration-sdk-core';
 
 import { Entities } from '../constants';
-import { SecureWorkloadScope, SecureWorkloadUser } from '../../types';
+import { SecureWorkloadUser } from '../../types';
 
 export function createUserEntity(user: SecureWorkloadUser): Entity {
   return createIntegrationEntity({
@@ -33,27 +33,6 @@ export function createUserEntity(user: SecureWorkloadUser): Entity {
   });
 }
 
-export function createScopeEntity(scope: SecureWorkloadScope): Entity {
-  return createIntegrationEntity({
-    entityData: {
-      source: scope,
-      assign: {
-        _type: Entities.SCOPE._type,
-        _class: Entities.SCOPE._class,
-        _key: scope.id,
-        id: scope.id,
-        shortName: scope.short_name,
-        name: scope.name,
-        description: scope.description || '', // Must be string by entity schema
-        vrfID: scope.vrf_id,
-        parentAppScopeID: scope.parent_app_scope_id,
-        childAppScopeIDs: scope.child_app_scope_ids,
-        displayName: scope.short_name,
-      },
-    },
-  });
-}
-
 export function createAccountUserRelationship(
   account: Entity,
   user: Entity,
@@ -61,28 +40,6 @@ export function createAccountUserRelationship(
   return createDirectRelationship({
     _class: RelationshipClass.HAS,
     from: account,
-    to: user,
-  });
-}
-
-export function createScopeRelationship(
-  scope: Entity,
-  childScope: Entity,
-): Relationship {
-  return createDirectRelationship({
-    _class: RelationshipClass.HAS,
-    from: scope,
-    to: childScope,
-  });
-}
-
-export function createScopeUserRelationship(
-  scope: Entity,
-  user: Entity,
-): Relationship {
-  return createDirectRelationship({
-    _class: RelationshipClass.ASSIGNED,
-    from: scope,
     to: user,
   });
 }
