+++ src/steps/index.ts
@@ -1,6 +1,7 @@
 import { accessSteps } from './access';
 import { accountSteps } from './account';
+import { scopeSteps } from './scope';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [...accountSteps, ...accessSteps, ...scopeSteps];
 
 export { integrationSteps };
